+++ samples/hello/hello_list_multi_selection.rb
@@ -1,5 +1,5 @@
 # Copyright (c) 2007-2020 Andy Maleh
-# 
+#
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
 # "Software"), to deal in the Software without restriction, including
@@ -7,10 +7,10 @@
 # distribute, sublicense, and/or sell copies of the Software, and to
 # permit persons to whom the Software is furnished to do so, subject to
 # the following conditions:
-# 
+#
 # The above copyright notice and this permission notice shall be
 # included in all copies or substantial portions of the Software.
-# 
+#
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 # EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
 # MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -24,22 +24,27 @@ class HelloListMultiSelection
     attr_accessor :provinces, :provinces_options
   
     def initialize
-      self.provinces_options=[
-        "",
-        "Quebec",
-        "Ontario",
-        "Manitoba",
-        "Saskatchewan",
-        "Alberta",
-        "British Columbia",
-        "Nova Skotia",
-        "Newfoundland"
+      self.provinces_options = [
+        '',
+        'Alberta',
+        'British Columbia',
+        'Manitoba',
+        'New Brunswick',
+        'Newfoundland and Labrador',
+        'Northwest Territories',
+        'Nova Scotia',
+        'Nunavut',
+        'Ontario',
+        'Prince Edward Island',
+        'Quebec',
+        'Saskatchewan',
+        'Yukon'
       ]
-      self.provinces = ["Quebec", "Manitoba", "Alberta"]
+      reset_provinces
     end
   
     def reset_provinces
-      self.provinces = ["Quebec", "Manitoba", "Alberta"]
+      self.provinces = ['Quebec', 'Manitoba', 'Alberta']
     end
   end
   
@@ -49,7 +54,7 @@ class HelloListMultiSelection
     person = Person.new
     
     shell {
-      grid_layout      
+      grid_layout
       
       text 'Hello, List Multi Selection!'
       
@@ -58,7 +63,7 @@ class HelloListMultiSelection
       }
       
       button {
-        text "Reset Selection To Defaults"
+        text 'Reset Selections To Default Values'
         
         on_widget_selected { person.reset_provinces }
       }